requirements.txt
fastapi
uvicorn
yt-dlp
openai-whisper
faster-whisper
python-multipart
